feat: support nest watermark
