+++ components/anchor/Anchor.tsx
@@ -1,5 +1,5 @@
 import classNames from 'classnames';
-import useEvent from 'rc-util/lib/hooks/useEvent';
+import { useEvent } from 'rc-util';
 import * as React from 'react';
 import scrollIntoView from 'scroll-into-view-if-needed';
