+++ components/watermark/context.ts
@@ -1,3 +1,4 @@
+import { useEvent } from 'rc-util';
 import * as React from 'react';
 
 export interface WatermarkContextProps {
@@ -11,5 +12,22 @@ const WatermarkContext = React.createContext<WatermarkContextProps>({
   add: voidFunc,
   remove: voidFunc,
 });
+
+export function usePanelRef(panelSelector?: string) {
+  const watermark = React.useContext(WatermarkContext);
+
+  const panelEleRef = React.useRef<HTMLElement>();
+  const panelRef = useEvent((ele: HTMLElement | null) => {
+    if (ele) {
+      const innerContentEle = panelSelector ? ele.querySelector<HTMLElement>(panelSelector)! : ele;
+      watermark.add(innerContentEle);
+      panelEleRef.current = innerContentEle;
+    } else {
+      watermark.remove(panelEleRef.current!);
+    }
+  });
+
+  return panelRef;
+}
 
 export default WatermarkContext;
