+++ package.json
@@ -127,7 +127,7 @@
     "rc-checkbox": "~3.1.0",
     "rc-collapse": "~3.7.0",
     "rc-dialog": "~9.2.0",
-    "rc-drawer": "~6.2.0",
+    "rc-drawer": "~6.4.1",
     "rc-dropdown": "~4.1.0",
     "rc-field-form": "~1.36.0",
     "rc-image": "~7.1.0",
